% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_functions.R
\name{var_distro}
\alias{var_distro}
\title{Generate Random Distribution}
\usage{
var_distro(n, mean, sd = 0, lower = -Inf, upper = Inf, method = "RND",
  lhs = NULL, distro = "normal", seed = 1234)
}
\arguments{
\item{n}{number of realizations}

\item{mean}{mean}

\item{sd}{standard deviation}

\item{lower}{lower limit of the truncated distribution}

\item{upper}{upper limit of the truncated distribution}

\item{method}{sampling method, either "RND" or "LHS". Default = "RND"}

\item{lhs}{vector of quatiles (obtained from LH sampling functions). Not required
if \code{method = "RND"}. Default = NULL}

\item{distro}{distribution}

\item{seed}{integer serving as seed number, work only for the RND option}
}
\value{
an array with n random values for the given parameters
}
\description{
Generate Random Distribution
}
\examples{
\dontrun{
distribution_1 <- var_distro(100, 10, 4)
hist(distribution_1)

distribution_2 <- var_distro(100, 10, 4, "truncated normal", 7, 13, seed = 123)
hist(distribution_2)
}

}
